I got it to run.
```
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 237 and 763 out of 1000.
The time was 00:00:01.3643186 for: iteration 0
Starting timing for: iteration 1
Winrate is 229 and 771 out of 1000.
The time was 00:00:01.2942737 for: iteration 1
Starting timing for: iteration 2
Winrate is 259 and 741 out of 1000.
The time was 00:00:01.3571262 for: iteration 2
Starting timing for: iteration 3
Winrate is 233 and 767 out of 1000.
The time was 00:00:01.3084063 for: iteration 3
Starting timing for: iteration 4
Winrate is 233 and 767 out of 1000.
The time was 00:00:01.3119005 for: iteration 4
Starting timing for: iteration 5
Winrate is 275 and 725 out of 1000.
The time was 00:00:01.2916204 for: iteration 5
Starting timing for: iteration 6
Winrate is 223 and 777 out of 1000.
The time was 00:00:01.3201651 for: iteration 6
Starting timing for: iteration 7
Winrate is 222 and 778 out of 1000.
The time was 00:00:01.2905296 for: iteration 7
Starting timing for: iteration 8
Winrate is 273 and 727 out of 1000.
The time was 00:00:01.2663081 for: iteration 8
Starting timing for: iteration 9
Winrate is 252 and 748 out of 1000.
The time was 00:00:01.3101002 for: iteration 9
Starting timing for: iteration 10
Winrate is 232 and 768 out of 1000.
The time was 00:00:01.3056029 for: iteration 10
Starting timing for: iteration 11
Winrate is 242 and 758 out of 1000.
The time was 00:00:01.3442653 for: iteration 11
Starting timing for: iteration 12
Winrate is 225 and 775 out of 1000.
The time was 00:00:01.2942728 for: iteration 12
Starting timing for: iteration 13
Winrate is 276 and 724 out of 1000.
The time was 00:00:01.2583389 for: iteration 13
Starting timing for: iteration 14
Winrate is 226 and 774 out of 1000.
The time was 00:00:01.2531166 for: iteration 14
```

Didn't I have this exact same error yesterday. What the hell is it now?