I did the redesigning that I wanted to do. I'll wind down for a bit here. Now the next thing to do is really the AC player.
I think this is pretty good. I have plenty of reason to be pleased by the current design of the library. It is very minimalistic. This particular design will make it very easy to use not just for poker, but any sort of game without much change. It will also make it easy to try out new architectures.

Insufficient designs tend to metastasize and spread like cancer, while good designs are foundational and steady as a rock.

Pretty soon I will have a good and working poker player. By the end of next month I will have them doing asynchronous training against themselves in no limit poker.

When I get to that stage, I will have power, real power.

Past that point it will just be a matter of chiseling away. I remember as a kid I used to use those game trainers. Breaking the game is the truest to their purpose. It is the way I am playing now that is bullshit.

I will make a full circle and go back in time, this time with my agents taking the place of cheats. I will just keep getting stronger and stronger without bound until the time of the Singularity.

A man's life is full of mistakes. He makes a mistake, swears not to make it again and makes a very similar one soon afterwards. The reason why I am here and now is because I could never break that cycle of imperfection. Computers have the power to effect that change. They are the true tabula rasa. They can be anything and become anything. I greatly envy that power and want to make it my own. The tragedy of my life is that imagination is not enough, but I certainly have perseverance.

I cannot even begin to describe how depressed I am that at this age I am much the same as I was two decades ago. Nobody could show me the way so I had to find it myself.