The tests are done and while not sharing features is unfortunately better than sharing features, it is still nowhere good enough to meet my requirements.
After a more thorough investigation, despite my initial impression there does not seem to be much difference between the Zap and non-Zap critic.

I am honestly disappointed by this.

Let me run the MC version so I have a thorough comparison. After that I'll put the PRONG layers back in. After that, I guess I'll go into the last step of the plan and complete the tour.

I am talking about implementing Rudder's reward composition. Using that to replace both the TD(0) and the MC will cut the variance down to nothing.

The theory of it is much better than for either TD learning or Monte Carlo, but it will commit me to using BPTT for the time being.

Well, for poker I will be able to deal with it.

Let me start out by running MC while I work on bringing back PRONG on the side. This should not be hard, though I am not expecting much at this point.

Well, I'll take it easy today. It is okay if I spend a day running tests occasionally.