Let me stop the test I am running since as the learning rates increase the winrate goes down. I was wrong, the critic with the steady state matrix is definitely better than the other one.
They both seem worse that Monte Carlo though so that is hardly a ringing endorsement.

Let me try one more test. I think I need to try even lower learning rates for the critic. I think the issue here is that I am not normalizing the rewards for the critic unlike for MC where PRONG is enough to do it on its own.

I am thinking that I could extend PRONG so it centers the inputs and the gradients. I think if I do it directly and only on the backward pass I could avoid needing to do the reprojection step for the bias.

Right now I am in thought as to what my next moves should be.

Let me just run the test.