The tests are done. Honestly, no. MC + PRONG beats AC + PRONG no matter what I do. In fact the difference is more pronounced since having a critic gives me more hyperparameters to tune. I have enough data on this now.
...Right now I am looking over the code trying to think of something where I could have gone wrong with the critic, but nothing comes to mind. It is reasonably high quality code at this point. It might seem convoluted, but any mistakes would show up as type errors.

I probably do not have a mistake. TD(0) simply works poorly. Since the purpose of a critic is to center the gradients, I'll aim at different ways towards getting that sort of thing. For starters I will work at extending PRONG with centering operations.

Regardless how that works, after that I will move straight into implementing Rudder. I am going to need its reward composition after all. I guess I'll finally find a use for RNNs. Differentiable plasticity would not be good with integrated gradients, but MI should fit the bill well. I'll implement DP along the way and use it for the actor as that architecture seems to have come into existence just for the sake of RL.

I'll leave Zap in the library for a while and before I remove it, put it in some text file and on a separate branch just in case I need it later.

Today was sure tiring. Waiting for those tests to finish was draining. It reminds me on waiting on stocks in my trading days. I am really bad at detaching. While I was not watching the winrates tick by, I was intending to rest during this time and instead spent it all on planning. I barely paid attention to the talk I was supposed to be watching.

Today was the first of many of these kinds of days. I'll be seeing them a lot more in the future. I've definitely reached a major milestone. I've surmounted the challenge and did not quite get the reward I expected, but I got a reward just the same.

PRONG makes standard PG stable enough that I can use it reliably.

Perhaps the greatest prize are those redesigns that I did. They will make it much easier for me to move through the design space going into the future.

Rudder definitely won't fail me. It is an intrinsic deep learning method with recurrent nets as a prerequisite. MC and TD can each be shot down with simple thought experiments, but at worst Rudder might at worst assign the reward to the wrong action. That is a much better outcome that the exponential variance and the instability of the other two methods.

The work that is to come will have to fall to the future me. The present day me did the best that he could. The day that is today is done.