I am not having much luck with epsilons lower that 2^-10. It optimizes, but it works a lot worse when the epsilon is higher.
There is no point in decreasing epsilon beyond a certain point it seems. The reason why such small epsilons are working is probably because I am setting the prong learning rate to be ** 0.85 of the normal one. This is a thing I picked up from the Zap paper and it is working absolutely brilliantly.

It is interesting that the right learning rates for online and batch version are within 2x of each other. There is no reason at all to rescale them it seems. Now that I have a better default I should try the PG agent again.