How annoying. I am still thinking about it.
As a method, what I have with PRONG is really brilliant. It just reprojects the gradients. But the critic on the other hand would hijack everything. Optimizing it would change the functionality of the actor.

Rudder - somehow I still do not want to go into it. Really, if I am doing a critic I should be using it, but...

It is a complicated method. In the paper they had to do a lot of hacks to get it to work.

Change of plans. What I am going to do tomorrow is do the Zap Q player.

I really do not get the feeling that I did everything I could with regards to Zap. The work is half finished at best.

Once I do the player I will know for sure the method's true worth and my soul will feel at peace. Taking a day or two to do this is not a big deal.

One more thing...maybe one of the issues that is making the critic unstable would be the updates during the backward chaining.

I'll know for sure once I really do it.

I should make a deep Zap Q player. If Zap is capable of stabilizing this then I know it will be able to handle the role of the critic.