```         inl learning_rate = {             actor=learning_rate             critic=learning_rate ** 0.85f32             shared=learning_rate             }
        inl prong learning_rate size =
            inl learning_rate = learning_rate ** 0.85f32
            inl steps_until_inverse_update = 128
            Feedforward.prong {learning_rate steps_until_inverse_update activation=Activation.linear; size initializer=Initializer.bias}

        inl actor = match d with {actor} -> Tuple.append (Tuple.wrap actor) (prong learning_rate.actor num_actions :: ()) | _ -> prong learning_rate.actor num_actions
        inl critic = match d with {critic} -> critic :: prong learning_rate.critic 1 :: () | _ -> prong learning_rate.critic 1
        inl shared = match d with {shared} -> shared | _ -> ()
```

Let me set the learning rates sensibly like this. I really like the ** 0.85 heuristic. It makes sense that the critic should train faster than the actor.

```
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 258 and 742 out of 1000.
The time was 00:00:03.2319965 for: iteration 0
Starting timing for: iteration 1
Winrate is 360 and 640 out of 1000.
The time was 00:00:04.3689022 for: iteration 1
Starting timing for: iteration 2
Winrate is 458 and 542 out of 1000.
The time was 00:00:05.2754467 for: iteration 2
Starting timing for: iteration 3
Winrate is 557 and 443 out of 1000.
The time was 00:00:05.6688704 for: iteration 3
Starting timing for: iteration 4
Winrate is 630 and 370 out of 1000.
The time was 00:00:05.4992300 for: iteration 4
Starting timing for: iteration 5
Winrate is 657 and 343 out of 1000.
The time was 00:00:05.8517244 for: iteration 5
Starting timing for: iteration 6
Winrate is 696 and 304 out of 1000.
The time was 00:00:05.6373202 for: iteration 6
Starting timing for: iteration 7
Winrate is 728 and 272 out of 1000.
The time was 00:00:05.1145977 for: iteration 7
Starting timing for: iteration 8
Winrate is 702 and 298 out of 1000.
The time was 00:00:05.3136507 for: iteration 8
Starting timing for: iteration 9
Winrate is 704 and 296 out of 1000.
The time was 00:00:05.6288170 for: iteration 9
Starting timing for: iteration 10
Winrate is 723 and 277 out of 1000.
The time was 00:00:04.9614962 for: iteration 10
Starting timing for: iteration 11
Winrate is 729 and 271 out of 1000.
The time was 00:00:05.1800824 for: iteration 11
Starting timing for: iteration 12
Winrate is 694 and 306 out of 1000.
The time was 00:00:05.2067482 for: iteration 12
Starting timing for: iteration 13
Winrate is 721 and 279 out of 1000.
The time was 00:00:05.3088038 for: iteration 13
Starting timing for: iteration 14
Winrate is 720 and 280 out of 1000.
The time was 00:00:05.1358432 for: iteration 14
```

At any rate, here is the actor with just PGs.

```
Starting timing for: iteration 0
Winrate is 296 and 704 out of 1000.
The time was 00:00:04.0164809 for: iteration 0
Starting timing for: iteration 1
Winrate is 460 and 540 out of 1000.
The time was 00:00:05.6469219 for: iteration 1
Starting timing for: iteration 2
Winrate is 545 and 455 out of 1000.
The time was 00:00:05.9846135 for: iteration 2
Starting timing for: iteration 3
Winrate is 609 and 391 out of 1000.
The time was 00:00:05.7341687 for: iteration 3
Starting timing for: iteration 4
Winrate is 665 and 335 out of 1000.
The time was 00:00:05.9033927 for: iteration 4
Starting timing for: iteration 5
Winrate is 673 and 327 out of 1000.
The time was 00:00:05.6815729 for: iteration 5
Starting timing for: iteration 6
Winrate is 698 and 302 out of 1000.
The time was 00:00:05.7545598 for: iteration 6
Starting timing for: iteration 7
Winrate is 744 and 256 out of 1000.
The time was 00:00:05.6350905 for: iteration 7
Starting timing for: iteration 8
Winrate is 708 and 292 out of 1000.
The time was 00:00:05.6938537 for: iteration 8
Starting timing for: iteration 9
Winrate is 715 and 285 out of 1000.
The time was 00:00:06.1212452 for: iteration 9
Starting timing for: iteration 10
Winrate is 745 and 255 out of 1000.
The time was 00:00:05.7063520 for: iteration 10
Starting timing for: iteration 11
Winrate is 734 and 266 out of 1000.
The time was 00:00:05.7784489 for: iteration 11
Starting timing for: iteration 12
Winrate is 725 and 275 out of 1000.
The time was 00:00:05.7976007 for: iteration 12
Starting timing for: iteration 13
Winrate is 728 and 272 out of 1000.
The time was 00:00:06.1646974 for: iteration 13
Starting timing for: iteration 14
Winrate is 737 and 263 out of 1000.
The time was 00:00:05.8531688 for: iteration 14
```

Oh, it works quite well, though I had to increase the learning rate to -10. Let me do that again.

```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 359 and 641 out of 1000.
The time was 00:00:05.0017902 for: iteration 0
Starting timing for: iteration 1
Winrate is 588 and 412 out of 1000.
The time was 00:00:05.9776473 for: iteration 1
Starting timing for: iteration 2
Winrate is 666 and 334 out of 1000.
The time was 00:00:06.0177879 for: iteration 2
Starting timing for: iteration 3
Winrate is 720 and 280 out of 1000.
The time was 00:00:05.9143559 for: iteration 3
Starting timing for: iteration 4
Winrate is 689 and 311 out of 1000.
The time was 00:00:05.7606905 for: iteration 4
Starting timing for: iteration 5
Winrate is 736 and 264 out of 1000.
The time was 00:00:05.5324650 for: iteration 5
Starting timing for: iteration 6
Winrate is 726 and 274 out of 1000.
The time was 00:00:05.5279760 for: iteration 6
Starting timing for: iteration 7
Winrate is 738 and 262 out of 1000.
The time was 00:00:05.3141337 for: iteration 7
Starting timing for: iteration 8
Winrate is 686 and 314 out of 1000.
The time was 00:00:05.4348897 for: iteration 8
Starting timing for: iteration 9
Winrate is 743 and 257 out of 1000.
The time was 00:00:05.7790188 for: iteration 9
Starting timing for: iteration 10
Winrate is 751 and 249 out of 1000.
The time was 00:00:05.9934548 for: iteration 10
Starting timing for: iteration 11
Winrate is 743 and 257 out of 1000.
The time was 00:00:05.7334832 for: iteration 11
Starting timing for: iteration 12
Winrate is 739 and 261 out of 1000.
The time was 00:00:05.9983427 for: iteration 12
Starting timing for: iteration 13
Winrate is 735 and 265 out of 1000.
The time was 00:00:05.8170476 for: iteration 13
Starting timing for: iteration 14
Winrate is 733 and 267 out of 1000.
The time was 00:00:05.5300428 for: iteration 14
```

Bingo. Let me try giong to -8.