It is visibly worse as expected.
```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 357 and 643 out of 1000.
The time was 00:00:04.3156200 for: iteration 0
Starting timing for: iteration 1
Winrate is 524 and 476 out of 1000.
The time was 00:00:05.2425157 for: iteration 1
Starting timing for: iteration 2
Winrate is 625 and 375 out of 1000.
The time was 00:00:05.7213575 for: iteration 2
Starting timing for: iteration 3
Winrate is 652 and 348 out of 1000.
The time was 00:00:05.6407706 for: iteration 3
Starting timing for: iteration 4
Winrate is 665 and 335 out of 1000.
The time was 00:00:05.5643226 for: iteration 4
Starting timing for: iteration 5
Winrate is 693 and 307 out of 1000.
The time was 00:00:05.4776796 for: iteration 5
Starting timing for: iteration 6
Winrate is 706 and 294 out of 1000.
The time was 00:00:05.9697410 for: iteration 6
Starting timing for: iteration 7
Winrate is 727 and 273 out of 1000.
The time was 00:00:05.3316426 for: iteration 7
Starting timing for: iteration 8
Winrate is 696 and 304 out of 1000.
The time was 00:00:05.0806540 for: iteration 8
Starting timing for: iteration 9
Winrate is 695 and 305 out of 1000.
The time was 00:00:05.4110011 for: iteration 9
Starting timing for: iteration 10
Winrate is 682 and 318 out of 1000.
The time was 00:00:05.2923277 for: iteration 10
Starting timing for: iteration 11
Winrate is 710 and 290 out of 1000.
The time was 00:00:04.9868612 for: iteration 11
Starting timing for: iteration 12
Winrate is 701 and 299 out of 1000.
The time was 00:00:05.1872605 for: iteration 12
Starting timing for: iteration 13
Winrate is 675 and 325 out of 1000.
The time was 00:00:04.8877007 for: iteration 13
Starting timing for: iteration 14
Winrate is 666 and 334 out of 1000.
The time was 00:00:05.0515088 for: iteration 14
```

Let me try changing the heuristic for the critic to `** 0.75`.