``` The learning_rate is 2 ** -9 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 359 and 641 out of 1000. The time was 00:00:05.0967790 for: iteration 0 Starting timing for: iteration 1 Winrate is 588 and 412 out of 1000. The time was 00:00:06.0273156 for: iteration 1 Starting timing for: iteration 2 Winrate is 666 and 334 out of 1000. The time was 00:00:05.9712186 for: iteration 2 Starting timing for: iteration 3 Winrate is 720 and 280 out of 1000. The time was 00:00:06.0441680 for: iteration 3 Starting timing for: iteration 4 Winrate is 689 and 311 out of 1000. The time was 00:00:05.9088791 for: iteration 4 Starting timing for: iteration 5 Winrate is 736 and 264 out of 1000. The time was 00:00:05.6174152 for: iteration 5 Starting timing for: iteration 6 Winrate is 726 and 274 out of 1000. The time was 00:00:05.6151749 for: iteration 6 Starting timing for: iteration 7 Winrate is 738 and 262 out of 1000. The time was 00:00:05.4708878 for: iteration 7 Starting timing for: iteration 8 Winrate is 686 and 314 out of 1000. The time was 00:00:05.7085556 for: iteration 8 Starting timing for: iteration 9 Winrate is 743 and 257 out of 1000. The time was 00:00:06.1259186 for: iteration 9 Starting timing for: iteration 10 Winrate is 751 and 249 out of 1000. The time was 00:00:06.2534067 for: iteration 10 Starting timing for: iteration 11 Winrate is 743 and 257 out of 1000. The time was 00:00:05.9857797 for: iteration 11 Starting timing for: iteration 12 Winrate is 739 and 261 out of 1000. The time was 00:00:06.1987334 for: iteration 12 Starting timing for: iteration 13 Winrate is 735 and 265 out of 1000. The time was 00:00:05.9838941 for: iteration 13 Starting timing for: iteration 14 Winrate is 733 and 267 out of 1000. The time was 00:00:05.6664256 for: iteration 14 Starting timing for: iteration 15 Winrate is 745 and 255 out of 1000. The time was 00:00:05.7763643 for: iteration 15 Starting timing for: iteration 16 Winrate is 729 and 271 out of 1000. The time was 00:00:05.2295758 for: iteration 16 Starting timing for: iteration 17 Winrate is 732 and 268 out of 1000. The time was 00:00:05.3626040 for: iteration 17 Starting timing for: iteration 18 Winrate is 726 and 274 out of 1000. The time was 00:00:05.9370510 for: iteration 18 Starting timing for: iteration 19 Winrate is 738 and 262 out of 1000. The time was 00:00:05.7228450 for: iteration 19 Starting timing for: iteration 20 Winrate is 701 and 299 out of 1000. The time was 00:00:05.9765551 for: iteration 20 Starting timing for: iteration 21 Winrate is 731 and 269 out of 1000. The time was 00:00:06.1165386 for: iteration 21 Starting timing for: iteration 22 Winrate is 719 and 281 out of 1000. The time was 00:00:06.1988198 for: iteration 22 Starting timing for: iteration 23 Winrate is 748 and 252 out of 1000. The time was 00:00:06.0694186 for: iteration 23 Starting timing for: iteration 24 Winrate is 733 and 267 out of 1000. The time was 00:00:05.4850350 for: iteration 24 Starting timing for: iteration 25 Winrate is 754 and 246 out of 1000. The time was 00:00:06.0755444 for: iteration 25 Starting timing for: iteration 26 Winrate is 738 and 262 out of 1000. The time was 00:00:05.3918502 for: iteration 26 Starting timing for: iteration 27 Winrate is 717 and 283 out of 1000. The time was 00:00:05.4253333 for: iteration 27 Starting timing for: iteration 28 Winrate is 702 and 298 out of 1000. The time was 00:00:05.4780298 for: iteration 28 Starting timing for: iteration 29 Winrate is 706 and 294 out of 1000. The time was 00:00:05.7462749 for: iteration 29 ```
In both cases in iteration 4 and 8 and the agents runs into some instability. Maybe there is a series of bad beats or something. The MC based agents can recover successfully from this while the TD one just dies.

Let me try adding another layer.

Edit:

```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 347 and 653 out of 1000.
The time was 00:00:05.9683037 for: iteration 0
Starting timing for: iteration 1
Winrate is 599 and 401 out of 1000.
The time was 00:00:07.2202618 for: iteration 1
Starting timing for: iteration 2
Winrate is 668 and 332 out of 1000.
The time was 00:00:06.7126249 for: iteration 2
Starting timing for: iteration 3
Winrate is 707 and 293 out of 1000.
The time was 00:00:06.7737185 for: iteration 3
Starting timing for: iteration 4
Winrate is 685 and 315 out of 1000.
The time was 00:00:06.6102417 for: iteration 4
Starting timing for: iteration 5
Winrate is 731 and 269 out of 1000.
The time was 00:00:06.7790724 for: iteration 5
Starting timing for: iteration 6
Winrate is 722 and 278 out of 1000.
The time was 00:00:07.1807559 for: iteration 6
Starting timing for: iteration 7
Winrate is 754 and 246 out of 1000.
The time was 00:00:06.8162520 for: iteration 7
Starting timing for: iteration 8
Winrate is 731 and 269 out of 1000.
The time was 00:00:06.9561824 for: iteration 8
Starting timing for: iteration 9
Winrate is 723 and 277 out of 1000.
The time was 00:00:06.9016540 for: iteration 9
Starting timing for: iteration 10
Winrate is 742 and 258 out of 1000.
The time was 00:00:06.9537542 for: iteration 10
Starting timing for: iteration 11
Winrate is 742 and 258 out of 1000.
The time was 00:00:06.9181286 for: iteration 11
Starting timing for: iteration 12
Winrate is 708 and 292 out of 1000.
The time was 00:00:06.8933352 for: iteration 12
Starting timing for: iteration 13
Winrate is 727 and 273 out of 1000.
The time was 00:00:06.9475535 for: iteration 13
Starting timing for: iteration 14
Winrate is 724 and 276 out of 1000.
The time was 00:00:06.6895805 for: iteration 14
Starting timing for: iteration 15
Winrate is 726 and 274 out of 1000.
The time was 00:00:06.9267124 for: iteration 15
Starting timing for: iteration 16
Winrate is 749 and 251 out of 1000.
The time was 00:00:06.8303332 for: iteration 16
Starting timing for: iteration 17
Winrate is 746 and 254 out of 1000.
The time was 00:00:06.9549101 for: iteration 17
Starting timing for: iteration 18
Winrate is 766 and 234 out of 1000.
The time was 00:00:06.6884994 for: iteration 18
Starting timing for: iteration 19
Winrate is 741 and 259 out of 1000.
The time was 00:00:06.8647165 for: iteration 19
Starting timing for: iteration 20
Winrate is 720 and 280 out of 1000.
The time was 00:00:06.5849125 for: iteration 20
Starting timing for: iteration 21
Winrate is 723 and 277 out of 1000.
The time was 00:00:07.2904250 for: iteration 21
Starting timing for: iteration 22
Winrate is 743 and 257 out of 1000.
The time was 00:00:06.7754781 for: iteration 22
Starting timing for: iteration 23
Winrate is 738 and 262 out of 1000.
The time was 00:00:07.3310073 for: iteration 23
Starting timing for: iteration 24
Winrate is 762 and 238 out of 1000.
The time was 00:00:06.7439781 for: iteration 24
Starting timing for: iteration 25
Winrate is 686 and 314 out of 1000.
The time was 00:00:06.9771589 for: iteration 25
Starting timing for: iteration 26
Winrate is 741 and 259 out of 1000.
The time was 00:00:06.6646121 for: iteration 26
Starting timing for: iteration 27
Winrate is 747 and 253 out of 1000.
The time was 00:00:06.9049988 for: iteration 27
Starting timing for: iteration 28
Winrate is 724 and 276 out of 1000.
The time was 00:00:06.9035168 for: iteration 28
Starting timing for: iteration 29
Winrate is 726 and 274 out of 1000.
The time was 00:00:07.0082992 for: iteration 29
```

Yeah, this is quite good. Adding another layer does not make things any worse at all. I guess I'll try removing all the layers next.