Here is the linear layer.
```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 319 and 681 out of 1000.
The time was 00:00:03.7168092 for: iteration 0
Starting timing for: iteration 1
Winrate is 506 and 494 out of 1000.
The time was 00:00:05.0179041 for: iteration 1
Starting timing for: iteration 2
Winrate is 644 and 356 out of 1000.
The time was 00:00:05.6814604 for: iteration 2
Starting timing for: iteration 3
Winrate is 719 and 281 out of 1000.
The time was 00:00:04.9416936 for: iteration 3
Starting timing for: iteration 4
Winrate is 708 and 292 out of 1000.
The time was 00:00:04.5021819 for: iteration 4
Starting timing for: iteration 5
Winrate is 713 and 287 out of 1000.
The time was 00:00:04.4734103 for: iteration 5
Starting timing for: iteration 6
Winrate is 748 and 252 out of 1000.
The time was 00:00:04.0261876 for: iteration 6
Starting timing for: iteration 7
Winrate is 704 and 296 out of 1000.
The time was 00:00:04.4259757 for: iteration 7
Starting timing for: iteration 8
Winrate is 706 and 294 out of 1000.
The time was 00:00:04.5608886 for: iteration 8
Starting timing for: iteration 9
Winrate is 739 and 261 out of 1000.
The time was 00:00:04.2952168 for: iteration 9
Starting timing for: iteration 10
Winrate is 737 and 263 out of 1000.
The time was 00:00:04.4376514 for: iteration 10
Starting timing for: iteration 11
Winrate is 705 and 295 out of 1000.
The time was 00:00:04.3912448 for: iteration 11
Starting timing for: iteration 12
Winrate is 727 and 273 out of 1000.
The time was 00:00:04.4624804 for: iteration 12
Starting timing for: iteration 13
Winrate is 731 and 269 out of 1000.
The time was 00:00:04.4674386 for: iteration 13
Starting timing for: iteration 14
Winrate is 745 and 255 out of 1000.
The time was 00:00:04.4995139 for: iteration 14
Starting timing for: iteration 15
Winrate is 725 and 275 out of 1000.
The time was 00:00:04.3492562 for: iteration 15
Starting timing for: iteration 16
Winrate is 720 and 280 out of 1000.
The time was 00:00:04.1957305 for: iteration 16
Starting timing for: iteration 17
Winrate is 741 and 259 out of 1000.
The time was 00:00:04.2376333 for: iteration 17
Starting timing for: iteration 18
Winrate is 725 and 275 out of 1000.
The time was 00:00:04.3571164 for: iteration 18
Starting timing for: iteration 19
Winrate is 694 and 306 out of 1000.
The time was 00:00:04.1277734 for: iteration 19
Starting timing for: iteration 20
Winrate is 725 and 275 out of 1000.
The time was 00:00:04.2527430 for: iteration 20
Starting timing for: iteration 21
Winrate is 729 and 271 out of 1000.
The time was 00:00:04.3459211 for: iteration 21
Starting timing for: iteration 22
Winrate is 721 and 279 out of 1000.
The time was 00:00:04.4674251 for: iteration 22
Starting timing for: iteration 23
Winrate is 744 and 256 out of 1000.
The time was 00:00:04.3344617 for: iteration 23
Starting timing for: iteration 24
Winrate is 722 and 278 out of 1000.
The time was 00:00:04.2162211 for: iteration 24
Starting timing for: iteration 25
Winrate is 748 and 252 out of 1000.
The time was 00:00:04.1076367 for: iteration 25
Starting timing for: iteration 26
Winrate is 745 and 255 out of 1000.
The time was 00:00:04.2132813 for: iteration 26
Starting timing for: iteration 27
Winrate is 717 and 283 out of 1000.
The time was 00:00:04.1718752 for: iteration 27
Starting timing for: iteration 28
Winrate is 712 and 288 out of 1000.
The time was 00:00:04.1441431 for: iteration 28
Starting timing for: iteration 29
Winrate is 707 and 293 out of 1000.
The time was 00:00:04.2497776 for: iteration 29
```

It works quite fine.

Let me wrap up here. Unlike what I said yesterday, I won't do Zap Q. Seeing MC work so well has convinced that doing that would be a waste of time. It seems that TD learning is simply too unstable to be worth bothering with in practice. It does not matter, it is not my problem.

By implementing the MC critic and using Zap's `** 0.85` heuristic to determine its learning rate, I've definitely benefited greatly. I got another level of improvement in stability and speed of learning over raw PG. Between the PRONG layers and this I am now actually in a very good position to make further strides.

I should make a thorough test to ensure that it is stable across multiple random seeds, but I think I won't be disappointed after I run this test. I fully expect it to succeed. Let me do this and I will wrap up for the day. I feel like taking a rest here.

I am at the point where I should seriously consider letting the agents train using self play.

But before that I definitely want to bring RNNs. If I can make them stable then I will really have everything. I haven't actually been able to do this last time I tried them out, but now I have prong and critics so things should go differently.