I am thinking right now. I've removed Zap from the library as it was making too much of a mess. What I will do is aim to make specialized MC and TD layers and then in case of Zap simply add its whitening step during the optimization pass in the generic TD layer. There is no point in doing it on the backward pass like with K-FAC since its weights are as large as the inputs. I am better off saving myself computation and doing it at the end in its case.
I can't believe I am still going at it. Aren't I supposed to be tired?

At any rate, I am not sure I will be able to finish this today, let alone test it. This is going to take me a day or two to get all in order. Making well designed, generic code requires a different mindset that simply putting a bunch of pieces together.

I have plenty of room to improve the existing implementations.

The way TD will be done will not really be taking advantage of backward chaining, but it is pretty useless right now as it is. Even Zap cannot make it work well. Some gradient averaging will help it significantly.

...I got `learning1` to compile at least so not everything is broken, but the players definitely are.

How about I stop, for real this time? All the work that I did today seems to have come in sudden bursts to me, but 5:20pm is hardly the time to start work on an entirely new fragment. I'll deal with this tomorrow.

All the pieces are already there, so it won't be hard to clean this all up. After that comes the RNNs.

The MI-RNN won't be hard at all to bring back in. Differentiable plasticity will take me a while longer to do, but I've been meaning to do this for months now.

Once I have that I will be set. It is self play, full NL Holdem and asynchronous training.

After that it will finally be time to look outwards to putting these things to work in the real world.

That is the real deal. That is where the accomplishments and the merit lie.