I got up early today and have been chilling for a bit. Yesterday I completely forgot about running that test suite so I am doing it now. As expected the results are great. It is still running, but the first couple of seeds are what I expected they would be.
What I will do today is continue the clean up work from yesterday. Make PG, MC and Zap into their own layers that I can compose freely. Right now there is too much code bound in the players. What could be moved to the ML library should be. The rest will be moved to the players themselves who should only need State and Action and the rest should go into PokerPlayers who should not be dependent on the implementation details of the players, but merely make the interfaces for them.

This is a good time to do this sort of work. After I am done with this, putting in the RNNs will be an half an hour of work.

Well, let me just start. It is time for the rebuild.