I was more focused yesterday it seems. I was too tired then to make stray thoughts then.
Now that I have this much, what is the next step? The PG layer. Let me go for it.

...It is just...let me exponentiate the learning rate in PRONG by 0.85.

I will add the learning_rate_modifier as a hyperparam into every layer. Rather than hardcode it, it would be better to do it like this. Having a different lr in the critic would be good without having to resort to modifying it in the test.