"``` inl td s {discount_factor reward v' v} = ```
The thing that is going through my mind right now is...wouldn't it be great if this function was an activation?

10:35am. Also...

```
        inl f front =
            match w with
            | {$front={x with epsilon}} -> x
            | {$front=()} -> {epsilon=default_epsilon}
            | _ -> ()
```

I do not like this default. Rather I think if I am picking PRONG I'd want it to whiten by default.

```
        inl f front =
            match w with
            | {$front={x with epsilon}} -> x
            | {$front=()} -> ()
            | _ -> {epsilon=default_epsilon}
```

Yeah, this is how it should be.

Well, almost.

```
| _ -> {$front={epsilon=default_epsilon}}
```

Now it will do what I want it to."