```             inl ac_td w =                 inl layers = {pg=pg w; td=td w}                 inl inits = module_map (inl _ {init} -> init) layers                 inl applies = module_map (inl _ {apply} -> apply) layers                 {                 init=inl s -> module_map (inl _ x -> x s) inits                 optimize=inl {learning_rate weights} s -> module_map (inl _ x -> x {learning_rate} s) weights                 apply=inl {input weights} s ->                     inl {out bck=action_bck} = applies.pg {input weights=weights.pg} s                     inl td_bck = applies.td {input weights=weights.td} s                     {                     out                     bck=inl _ ->                         ()                     }                 } ```
Ah, what the hell am I doing? AC cannot be a layer. It must be a player. Also I forgot that TD is supposed to be only half PRONG. The other half should be Zap. Am I getting impatient here? Or am I falling in love so much with my own code that I am forgetting what it is supposed to be doing in the first place?

Yeah, I am almost done with this. Let stop here to eat, and I'll make a Zap TD variant.