I cannot really express the depth of my disappointment at this code. Yesterday, I was wondering how it was going so well for me despite me being so tired. Here I am wondering why I am producing turds despite being so fresh. I think I started out the day producing good code.
What the fuck is going on here?

I am beyond pissed at this.

...Hrrrgghhh...

Change of plans. Instead of spreading out the functionality and duplicating code all over the place, what I will do is push it all into prong. It will make the whiten function convoluted because of all the paths it could take, but at least it will all be in a single place.

I cannot believe I am getting beaten by this. Am I really going to have to spend an entire day on this? It seems so.

It is not code duplication that is the problem per se. Duplicate functionality on the other hand definitely is as the brain has a lot of difficulty remembering where everything is. The worst kind of bugs come from duplicating functionality.