This has to be right. But I am not done yet. For TD learning I am really missing what to do about the bias term. I need to go through the math for Zap.
I explained it succinctly to /u/tihokan. Let me dig it up and see what happens when I add a bias term to...

Ah, I get it. The TD learning algorithm is just the backward pass for the linear layer. So the gradients for the weight update will not be affected by the lack of bias term. I can update it normally.