The error of forgetting to optimize shared errors is really a heinous one. Damn it.
I'll have to repeat the tests. Forget that for now though.

I've redone the players, but I move on to them, I want to make Mnist work again.

Though it is 6:30pm I do not feel like stopping at all. Let me do this for as long as I want.

I just want to get testing out of the way so I can move to the fun stuff.