Ah, this code, this code...trully I have no outlet for my hatred except this work. It is really a pity is taking this long. The real world is such an eyesore.
How long will this last?

The answer is as always - until it is done. Even after I am done with Rudder and RNNs, I will still be drawing less that 1% on NNs true power.

But this stuff that I am working on here is what I wanted to be working on all the way back in 2015 and 2016. I am finally at the level of inception. The last few months were a long beginning.

Really, why are all programming languages so dogshit? It is not like it would have been difficult for somebody to invent something like Spiral, but for some reason it did not happen. Having to spent like 2.5 years starting from 2016 to get to this point really held me back quite a while.

I need to go through the paces properly from here.

Yesterday, I was mad as hell, but now I feel a bit determined instead.

I need to focus.

The next step would be to do the testing.

It is no wonder this code made me mad. I really want to slice it up into individual pieces, but the tools to abstract it properly simply aren't there. I need something more than Spiral offers me. I am not sure how that would even be possible.

So by necessity `whiten` will have to be what it is - heavily branching. This will be a pain to test, but let me do it. I think it should be correct in essence, but let get the type errors out of the way first.